+++ src/routes/calender/index.jsx
@@ -3,7 +3,7 @@ import { useState, useEffect, useCallback } from "react";
 import styled from "styled-components";
 import backIcon from "@icon/home/goBack.svg";
 import { ExerciseAPI } from "../../shared/lib/api";
-import { getExerciseProgressById, formatDateForStorage } from "../../shared/lib/exerciseProgress";
+import ExerciseCard from "./ExerciseCard";
 
 export default function ReportPage() {
   const navigate = useNavigate();
@@ -49,12 +49,6 @@ export default function ReportPage() {
     setSelectedDate(newDate);
   };
 
-  // 오늘 날짜인지 확인
-  const isToday = () => {
-    const today = new Date();
-    const selected = new Date(selectedDate);
-    return today.toDateString() === selected.toDateString();
-  };
 
   // 맞춤 운동 다시하기 버튼 클릭 핸들러
   const handleRetryExercise = (exercise) => {
@@ -66,30 +60,6 @@ export default function ReportPage() {
     });
   };
 
-  // 개별 운동 진행률 계산
-  const calculateExerciseProgress = (exercise) => {
-    const dateString = formatDateForStorage(selectedDate);
-    const savedProgress = getExerciseProgressById(exercise.exerciseId, dateString);
-
-    let currentReps = 0;
-    let percentage = 0;
-
-    if (savedProgress) {
-      if (savedProgress.isCompleted) {
-        // 운동이 완료된 경우
-        currentReps = exercise.reps;
-        percentage = 100;
-      } else {
-        // 운동이 진행 중인 경우
-        currentReps = savedProgress.currentCount || 0;
-        percentage = savedProgress.percentage || 0;
-      }
-    }
-
-    const totalReps = exercise.reps;
-
-    return { current: currentReps, total: totalReps, percentage };
-  };
 
   // API 요청 함수
   const fetchExercisesForDate = useCallback(
@@ -179,56 +149,12 @@ export default function ReportPage() {
         {!loading && !error && exercises.length > 0 && (
           <ExerciseList>
             {exercises.map((exercise) => (
-              <ExerciseCard key={exercise.exerciseId}>
-                <ExerciseName>{exercise.name}</ExerciseName>
-                <ExerciseReps>
-                  {exercise.reps}
-                  {exercise.unit}
-                </ExerciseReps>
-                <ExerciseAdvantages>{exercise.advantages}</ExerciseAdvantages>
-
-                {exercise.details && exercise.details.length > 0 && (
-                  <ExerciseDetails>
-                    {exercise.details.map((detail) => (
-                      <ExerciseDetailItem key={detail.id}>
-                        <ExerciseImage
-                          src={detail.imageUrl}
-                          alt={`${exercise.name} ${detail.id}`}
-                        />
-                        <ExerciseDescription>{detail.description}</ExerciseDescription>
-                      </ExerciseDetailItem>
-                    ))}
-                  </ExerciseDetails>
-                )}
-
-                {/* 오늘 날짜일 때만 각 운동별 진행률과 버튼 표시 */}
-                {isToday() && (
-                  <>
-                    <ExerciseProgressBar>
-                      <ProgressSection>
-                        <InitialCount>0</InitialCount>
-                        <ExerciseProgressText>
-                          <ExerciseCurrentProgress>
-                            {calculateExerciseProgress(exercise).current}
-                          </ExerciseCurrentProgress>
-                          <ExerciseProgressDivider>
-                            {" "}
-                            / {calculateExerciseProgress(exercise).total}
-                          </ExerciseProgressDivider>
-                        </ExerciseProgressText>
-                      </ProgressSection>
-                      <ExerciseProgressBarContainer>
-                        <ExerciseProgressBarFill
-                          percentage={calculateExerciseProgress(exercise).percentage}
-                        />
-                      </ExerciseProgressBarContainer>
-                    </ExerciseProgressBar>
-                    <ExerciseRetryButton onClick={() => handleRetryExercise(exercise)}>
-                      맞춤 운동 다시하기
-                    </ExerciseRetryButton>
-                  </>
-                )}
-              </ExerciseCard>
+              <ExerciseCard
+                key={exercise.exerciseId}
+                exercise={exercise}
+                selectedDate={selectedDate}
+                onClick={() => handleRetryExercise(exercise)}
+              />
             ))}
           </ExerciseList>
         )}
@@ -361,140 +287,3 @@ const ExerciseList = styled.div`
   gap: 16px;
 `;
 
-const ExerciseCard = styled.div`
-  background: white;
-  border-radius: 12px;
-  padding: 20px;
-  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
-`;
-
-const ExerciseName = styled.h3`
-  color: #1a1a1a;
-  font-family: Pretendard;
-  font-size: 18px;
-  font-weight: 600;
-  margin: 0 0 8px 0;
-`;
-
-const ExerciseReps = styled.p`
-  color: #427bf0;
-  font-family: Pretendard;
-  font-size: 16px;
-  font-weight: 600;
-  margin: 0 0 12px 0;
-`;
-
-const ExerciseAdvantages = styled.p`
-  color: #666;
-  font-family: Pretendard;
-  font-size: 14px;
-  line-height: 1.5;
-  margin: 0 0 16px 0;
-`;
-
-const ExerciseDetails = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-`;
-
-const ExerciseDetailItem = styled.div`
-  display: flex;
-  gap: 12px;
-  align-items: flex-start;
-`;
-
-const ExerciseImage = styled.img`
-  width: 80px;
-  height: 80px;
-  border-radius: 8px;
-  object-fit: cover;
-  flex-shrink: 0;
-`;
-
-const ExerciseDescription = styled.p`
-  color: #333;
-  font-family: Pretendard;
-  font-size: 14px;
-  line-height: 1.4;
-  margin: 0;
-  flex: 1;
-`;
-
-const ExerciseProgressBar = styled.div`
-  margin: 16px 0;
-`;
-
-const ExerciseProgressText = styled.div`
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-  margin-bottom: 8px;
-`;
-
-const ExerciseCurrentProgress = styled.span`
-  color: #427bf0;
-  font-family: Pretendard;
-  font-size: 35px;
-  font-weight: 700;
-`;
-
-const ExerciseProgressDivider = styled.span`
-  color: #959595;
-  font-family: Pretendard;
-  font-size: 20px;
-  font-weight: 600;
-`;
-
-const ExerciseProgressBarContainer = styled.div`
-  width: 100%;
-  height: 6px;
-  background-color: #f0f0f0;
-  border-radius: 3px;
-  overflow: hidden;
-`;
-
-const ExerciseProgressBarFill = styled.div`
-  width: ${(props) => props.percentage}%;
-  height: 100%;
-  background-color: #427bf0;
-  border-radius: 3px;
-  transition: width 0.3s ease;
-`;
-
-const ExerciseRetryButton = styled.button`
-  width: 100%;
-  background-color: #427bf0;
-  color: white;
-  border: none;
-  border-radius: 8px;
-  padding: 12px;
-  font-family: Pretendard;
-  font-size: 14px;
-  font-weight: 600;
-  cursor: pointer;
-  transition: background-color 0.2s ease;
-  margin-top: 12px;
-
-  &:hover {
-    background-color: #3366cc;
-  }
-
-  &:active {
-    background-color: #2952a3;
-  }
-`;
-
-const ProgressSection = styled.div`
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: space-between;
-`;
-
-const InitialCount = styled.span`
-  color: #959595;
-  font-family: Pretendard;
-  font-size: 20px;
-  font-weight: 600;
-`;
